prova di soluzione

+++ 2022-06-08/invsymlink.c
@@ -1,4 +1,4 @@
-include <limits.h>
+#include <limits.h>
 #include <stdlib.h>
 #include <stdio.h>
 #include <sys/types.h>
@@ -6,24 +6,28 @@ include <limits.h>
 #include <unistd.h>
 
 int main(int argc, char *argv[]) {
-    // 8:54
+    if (argc != 2) {
+        printf("usage: %s <symlink>\n", argv[0]);
+        exit(1);
+    }
     struct stat statbuf;
     int stat_ret = lstat(argv[1], &statbuf);
-    if (S_ISLNKstat_ret) {
-        // TODO:
+    if (stat_ret == -1) {
+        perror("lstat");
+        exit(1);
     }
 
-    if (!IS_
+    if (!S_ISLNK(statbuf.st_mode)) {
+        printf("not a symlink\n");
+        exit(1);
+    } 
 
     char *abspath = realpath(argv[1], NULL);
 
-    // TODO: check is symlink
     unlink(argv[1]);
     link(abspath, argv[1]);
     unlink(abspath);
     symlink(argv[1], abspath);
-
-    printf("%s\n", abspath);
 
     free(abspath);
 }
